Docker support docker_name
Add docker_name:

In the form, you can just input a name which is different from the other names of the same environment and version, you do not need to consider whether it is unique one in all the docker_names.

While deploying, it will automatically transform to $(env_name)_$(version)_$(deocker_name) in order to avoid duplicate names.